pyproject.toml
[build-system]
requires = ["setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trivial-graphics"
version = "0.0.4"
authors = [
  { name="George Watson", email="gigolo@hotmail.co.uk" },
]
description = "Pythonic OpenGL Bindings"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pyopengl",
    "pillow",
]
classifiers = [
    "Natural Language :: English",
    "Intended Audience :: Developers",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Programming Language :: Python",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
license = { text = "GPLv3" }

[project.urls]
Homepage = "https://github.com/takeiteasy/trivial-graphics"
Issues = "https://github.com/takeiteasy/trivial-graphics/issues"
